fix bugs on closing and save values of filter element
