+++ src/components/Filters/Filters.jsx
@@ -1,9 +1,5 @@
-import { useState, useEffect } from 'react';
-import { useDispatch } from 'react-redux';
-
 import FiltersByAge from 'components/FiltersByAge/FiltersByAge';
 import FiltersByGender from 'components/FiltersByGender/FiltersByGender';
-import { getNoticesByFilter } from 'redux/notices/noticesOperations';
 
 import {
   FilterWrapper,
@@ -12,17 +8,7 @@ import {
   TitleFilters,
 } from './Filters.styled';
 
-const Filters = () => {
-  const [filtersState, setFiltersState] = useState({});
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    if (Object.keys(filtersState).length === 0) {
-      return;
-    }
-    dispatch(getNoticesByFilter(filtersState));
-  }, [dispatch, filtersState]);
-
+const Filters = ({ setFiltersState }) => {
   return (
     <FilterWrapper>
       <TitleFilters>Filters</TitleFilters>
