--- src/components/FiltersByGender/FilterByGender.styled.js
@@ -1,1 +1,0 @@
-const { styled } = require('styled-components');
