+++ src/redux/notices/noticesSelectors.js
@@ -12,6 +12,8 @@ export const selectFilter = state => state.searchNotices;
 
 export const selectFavoriteNotices = state => state.notices.noticesFavorite;
 
+export const selectStatusFilters = state => state.notices.filters;
+
 // export const selectStatusFilter = state => state.filters.status;
 export const selectTotalNotices = state => state.notices.totalNotices;
